pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telescope-webapp"
version = "0.1.0"
description = "Web app for Cherenkov telescope DL3 analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pydantic>=2.9,<3.0",
  "pydantic-settings>=2.5,<3.0",
  "streamlit>=1.39,<2.0",
  "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "httpx>=0.27,<1.0",
  "ruff>=0.7,<1.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
pythonpath = ["backend", "."]
